+++ packages/create-gasket-app/test/unit/scaffold/config-builder.test.js
@@ -1,6 +1,6 @@
 /* eslint-disable max-statements */
-
-const ConfigBuilder = require('../../../lib/scaffold/config-builder');
+import { jest } from '@jest/globals';
+const { ConfigBuilder } = await import('../../../lib/scaffold/config-builder');
 
 const pluginOne = {
   name: 'gasket-plugin-one'