+++ packages/gasket-plugin-docusaurus/lib/create.js
@@ -1,6 +1,10 @@
-const { devDependencies } = require('../package.json');
+const { name, version, devDependencies } = require('../package.json');
 
-module.exports = async function create(gasket, { pkg }) {
+module.exports = function create(gasket, { pkg, gasketConfig }) {
+  gasketConfig.addPlugin('pluginDocusaurus', name);
+  pkg.add('dependencies', {
+    [name]: `^${version}`
+  });
   pkg.add('devDependencies', {
     '@docusaurus/core': devDependencies['@docusaurus/core'],
     '@docusaurus/preset-classic': devDependencies['@docusaurus/preset-classic'],