+++ packages/create-gasket-app/CHANGELOG.md
@@ -1,5 +1,25 @@
 # `create-gasket-app`
 
+
+
+- Add preset lifecycles `presetPrompt` and `presetConfig` ([#736])
+- ESM refactor, package is now type module
+- Gasket apps default to `type: module`
+- Update the `load-preset` functionality
+- Remove `addPlugins` utility
+- Change the implementation of `gasket` usage to `makeGasket`
+- Remove obsolete files
+- Remove references to `@gasket/resolve`
+- Remove `plugins`, `bootstrap` & `generate` flags
+- Adjust prompt flag to `no-prompt`
+- Remove plugin start references
+- Adjust default plugins to remove lifecycle, metadata & start plugins
+- Adjust `CreateContext` properties
+- Add several methods to the `ConfigBuilder`
+- Remove `setup.js` getopts script
+- Update `writeGasketConfig` to write in ESM
+- Update tests
+
 ### 6.46.6
 
 - Fix resolve cli path ([#712])
@@ -39,3 +59,4 @@
 [#442]: https://github.com/godaddy/gasket/pull/442
 [#460]: https://github.com/godaddy/gasket/pull/460
 [#712]: https://github.com/godaddy/gasket/pull/712
+[#736]: https://github.com/godaddy/gasket/pull/736