+++ packages/gasket-plugin-response-data/CHANGELOG.md
@@ -1,5 +1,10 @@
 # `@gasket/plugin-response-data`
 
+
+
+- Add plugin import to gasket file ([#736])
+- Add itself to the app package file
+
 ### 6.45.2
 
 - Add `cross-env`, adjust test script, remove `eslint-plugin-mocha`, refactor test ([#670])
@@ -63,3 +68,4 @@
 [#436]: https://github.com/godaddy/gasket/pull/436
 [#460]: https://github.com/godaddy/gasket/pull/460
 [#670]: https://github.com/godaddy/gasket/pull/670
+[#736]: https://github.com/godaddy/gasket/pull/736